projeto-genesis/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
langchain
chromadb
llama-cpp-python
